Fix bugs login

+++ app/Controllers/Users.php
@@ -39,6 +39,7 @@ class Users extends ResourceController
 
     public function update($id = null)
     {
+        if ($id == 1)  return $this->failUnauthorized('Access denied');
         if ($this->usersModel->find($id)) {
             if ($this->usersModel->update($id, $this->request->getVar())) {
                 return $this->respond(['message' => "Updated successfully"]);
@@ -52,6 +53,8 @@ class Users extends ResourceController
 
     public function delete($id = null)
     {
+        if($id==1)  return $this->failUnauthorized('Access denied');
+        
         if ($this->usersModel->find($id)) {
             if ($this->usersModel->delete($id)) {
                 return $this->respond(['message' => "Successfully deleted"]);
